i18n: localize monster info with issues to saves.
Localized monster name is written to save (again), and it's not good, since it's not been translated.
On the other hand, I can't figure out how to translate "minotaur simulacrum (sleeping)" and such out of context of monster description generation function.
It looks we want 2 funcs with same logic but that differ only in wheter the translation is done or not.