i18n: a bit more order in translations.
Figured out, that it is of no use to throw all the strings into crawl_locale.h file, because there are plenty of them that live in a single context and thus can happilu be collected with xgettext. So from now on, such strings will be left alone and not moved to the "registry". Only those which, like monsters, attack names and some others, can appear in multiple contexts, should be handled differently, because the count of their translations is the multiplication of their own count and their contexts count.